73: Added extra button type and rounded button type
